conda specific: null-terminate PREFIX so binary relocation works
Conda installer will attempt to detect all prefix strings in binaries and
change them to the new installed prefix.  However, it assumes the strings
are null-terminated, and FORTRAN strings aren't.  This patch makes PREFIX
null-terminated, so conda can successfully relocate the binary.  It also
adds a few utility functions to convert the null-terminated string to the
usual FORTRAN kind.