Reworked the build system
Makefiles have been reworked to correctly record dependencies and rebuild
code on an as-need basis.  Now it's also possible to run parallel builds,
i.e.  something like:

    make -j4

will work now, whereas it threw an error before. The default target `all`
will build all programs from `main/` and store the binaries into `bin/`. It
will also build the Python module, storing the library in `lib`. For
backwards compatibility copies of binaries/modules are still kept in `main/`
and `python/`.

Because of how Fortran interacts with make (poorly), it's very difficult to
correctly maintain a tree of Makefiles -- most have now been coalesced into
a single one (in `build/Makefile`).

All builds are now done in the `build/` directory. This makes it easier to
make Fortran (various compilers) and `make` play well together. A beneficial
side-effect is that source directories are not poluted by generated files.

The detailed dependency rules are stored in build/make.depends files. This
file is generated by running `make depends`, but we distribute it with the
source code as it doesn't have to be updated unless module USE-age is
changed or new files are added. See the documentation in `Makefile` for
details.

Tested to work with gfortran and ifort.